requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
reportlab==4.0.0
PyMuPDF==1.23.8
google-generativeai
werkzeug==2.3.7
